pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "beancountpilot"
version = "0.1.0"
description = "AI 增强的智能交易分类和工作流增强工具"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "BeancountPilot Contributors"}
]
keywords = [
    "beancount",
    "accounting",
    "finance",
    "ai",
    "llm",
    "classification",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial :: Accounting",
]

dependencies = [
    "fastapi>=0.128.0",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy>=2.0.36",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "openai>=1.58.0",
    "pyyaml>=6.0.2",
    "python-dotenv>=1.0.1",
    "pandas>=2.2.3",
    "requests>=2.32.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "black>=24.10.0",
    "ruff>=0.9.0",
    "mypy>=1.14.0",
]

frontend = [
    "streamlit>=1.53.0",
    "streamlit-extras>=0.5.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/BeancountPilot"
Documentation = "https://github.com/yourusername/BeancountPilot#readme"
Repository = "https://github.com/yourusername/BeancountPilot"
Issues = "https://github.com/yourusername/BeancountPilot/issues"

[project.scripts]
beancountpilot = "src.api.main:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
exclude = ["tests*"]

[tool.setuptools.package-data]
"beancountpilot" = ["config/*.yaml"]

# Black 配置
[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

# Ruff 配置
[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",
    "F",
    "I",
    "N",
    "W",
]
ignore = []

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

# MyPy 配置
[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

# Pytest 配置
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
